requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
faiss-cpu==1.13.2
filelock==3.20.3
fsspec==2026.1.0
google-auth==2.48.0
google-genai==1.60.0
gunicorn==25.0.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.3.0
python-docx==1.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
whitenoise==6.11.0
